+++ src/utils/format-bytes.js
@@ -1,0 +1,9 @@
+// https://gist.github.com/zentala/1e6f72438796d74531803cc3833c039c
+export default function formatBytes(bytes, decimals) {
+  if (bytes === 0) return '0 Bytes';
+  const k = 1024;
+  const dm = decimals || 2;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${parseFloat((bytes / (k ** i)).toFixed(dm))} ${sizes[i]}`;
+}