+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Modules with a 'file' type option now get a file selector/uploader instead of a text box 
+
 ## [2.2.0] - 2023-03-31
 
 -   Move Agent Tasks under the Agents page